chore(components): move Icon buttons to Buttons folder
